Mod links fix, message syntax improved
Fixed the moderation of links (I think)
Changed the raffle and poll messages to precise to only enter if you
haven't already.
If Twitch auto-censors the bad word you're trying to add, the bot warns
you about it.
Prepared the code for a !channel follow/unfollow command. But I gotta
rethink that, there's probably no point to it.